IOS Screen EditProfile + kotlin functions , map screen for ios

+++ shared/src/commonMain/kotlin/org/example/project/utils/extension/UserDtoExtension.kt
@@ -41,7 +41,7 @@ fun Dog.toDto(ownerId: String): DogDto = DogDto(
     id         = this.id,
     name       = this.dogName,
     breed      = this.dogBreed,
-    weight     = this.dogWeight.toInt(),
+    weight     = this.dogWeight,
     imgUrl     = this.dogPictureUrl ?: "",
     isFriendly = this.isFriendly,
     isMale     = (this.dogGender == Gender.MALE),
